Move image related functions from store.go to image.go
